+++ lib/features/devices/device/widgets/neon_brightness_tile/neon_brightness_tile_dialog/bloc/neon_brightness_tile_dialog_event.dart
@@ -8,7 +8,7 @@ abstract class NeonBrightnessTileDialogEvent extends Equatable {
 }
 
 class NewBrightness extends NeonBrightnessTileDialogEvent {
-  final double brightness;
+  final int brightness;
 
   const NewBrightness(this.brightness);
 }
